// node/interactive-analysis/packages/weavess/src/ts/components/waveform-display/components/axes/y-axis/types.ts
export interface YAxisProps {
  /** The height in percentage */
  heightInPercentage: number;

  /** Max amplitude as a number */
  maxAmplitude: number;

  /** Min amplitude as a number */
  minAmplitude: number;

  /** the y axis ticks */
  yAxisTicks?: number[];
}

// tslint:disable-next-line:no-empty-interface
export interface YAxisState {}